story(libs-components-ui): add example for form select helper
story(libs-components-ui): add example for form select helper

story(libs-components-ui): add example for form select helper
