+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,16 +1,16 @@
-<!-- Please make sure you have read the submission guidelines before posting an PR -->
-<!-- https://github.com/nrwl/nx/blob/master/CONTRIBUTING.md#-submitting-a-pr -->
-
 <!-- Please make sure that your commit message follows our format -->
-<!-- Example: `fix(nx): must begin with lowercase` -->
+<!-- Example: `fix(libs-ui-component): must begin with lowercase` -->
 
 ## Current Behavior
+
 <!-- This is the behavior we have today -->
 
 ## Expected Behavior
+
 <!-- This is the behavior we should expect with the changes in this PR -->
 
 ## Related Issue(s)
+
 <!-- Please link the issue being fixed so it gets closed when this is merged. -->
 
 Fixes #
