+++ libs/components/ui/src/components/select/Select.types.ts
@@ -1,9 +1,4 @@
 import { PropsFromKeys } from '@codelab/shared/interface/props'
-import {
-  ReactNodeI,
-  ReactNodeTypeEnum,
-  NodeTypeEnum,
-} from '@codelab/shared/interface/node'
 
 export namespace Select {
   export const propKeys = [
@@ -70,34 +65,4 @@ export namespace Select {
   export type Props = PropsFromKeys<typeof propKeys[number]>
 
   export type OptionProps = PropsFromKeys<typeof optionPropKeys[number]>
-
-  export const createSelect = (): ReactNodeI<Props> => ({
-    type: 'Select',
-    nodeType: 'React',
-    props: {
-      style: {
-        width: 120,
-      },
-    },
-  })
-
-  export const createOptions = (
-    type: typeof NodeTypeEnum,
-  ): Array<ReactNodeI<OptionProps>> =>
-    Object.entries(type).map(([key, value]: [string, any]) => ({
-      type: 'Select.Option',
-      nodeType: 'React',
-      props: {
-        value,
-      },
-      children: [
-        {
-          type: 'Text',
-          nodeType: 'React',
-          props: {
-            value: key,
-          },
-        },
-      ],
-    }))
 }
